+++ slidershow/slidershow.js
@@ -115,6 +115,13 @@ function get_menu() {
             <div id="hud-gps"></div>
             <div id="hud-tag"></div>
             <div id="hud-counter"></div>
+            <div id="hud-menu">
+                <button data-role="aux_window" title="Auxiliary window Alt+W">&#127916;</button>
+                <button data-role="tagging" title="Tagging mode Alt+T">&#128204;</button>
+                <button data-role="thumbnails" title="Thumbnails Alt+J">&#127895;</button>
+                <button data-role="export" title="Export Ctrl+S">&#128190;</button>
+                <button data-role="help" title="Help H">&#8505;</button>
+            </div>
         </div>
         <div id="hud-thumbnails"></div>
     </div>
@@ -125,11 +132,7 @@ function get_menu() {
         </div>
         <div id='start-wrapper'>
             Start presenting<br />
-            <button id="start">&#9654;</button><br />
-            <button data-role="aux_window" title="Auxiliary window Alt+W">&#127916;</button>
-            <button data-role="tagging" title="Tagging mode Alt+T">&#128204;</button>
-            <button data-role="export" title="Export Ctrl+S">&#128190;</button>
-            <button data-role="help" title="Help H">&#8505;</button>
+            <button id="start">&#9654;</button>
             </div>
 
         <div>