+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.8.0
 * auxiliary window to track the presentation pace
+* thumbnail ribbon
 
 ## 0.7.0
 Every important feature working as expected.